+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
+- reverse and reverse! for LowerTriangularArray, in longitude or latitude [#720](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/720)
 - Speed up CI for land models [#719](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/719)
-- Move documentation to SpeedyWeatherDocumentation repository [#713](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/713)
+- Move documentation to SpeedyWeatherDocumentation repository [#713](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/713), [#722](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/722), [#723](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/723), and [#725](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/725)
+- Ensure docs don't deploy with `.jld2` or `.nc` files [#724](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/724)
 
 ## v0.15